requirements.txt
newsapi-python
beautifulsoup4
pandas
nltk
spacy
scikit-learn
python-dotenv
requests
geopy
xgboost 